These patches fix bugs in the integrator in conjunction with algint.
algint uses the fluid variable intvar to store the integration variable.
A substition did not update the value of intvar and a subsequent call to
algebraiccase would give a completely incorrect result.
The patch to driver.red makes sure that the value of intvar is current.

The second patch handles a null value of cmatrix. This can occur when algint
calls the integrator recursively.  A segmentation fault in compiled PSL code
could occur.