Detect integer powers of exponentials in findzvars; this avoids a simplication error in some cases like
  int(e^(e^(3*x)),x)

or 

  z:=e^x;
  u:=sqrt(2)*z/(z^2+1);
  int(e^(u^6)/u^2,x);

(distilled from odesolve zimmer.tst case #7).

Bug fix for

  int(f^(b*x)*sec(c + d*x)^3,x);

in trcase.red: a temp variable was not reset and the value returned was
the sum of the unevaluated integral and the integrand.

Add substitution for exponential of (at the moment only negative) powers of
the integration variable. e.g

  int(e^(1/x^3)/x^7,x);

Add call to simpint1 to try pattern matching after a substitution of rational
power of integration var.